Initial net6.h
The initial net6.h just has the definition of an IPv6 address and IPv6
header. Subsequent changes will build on this.

Signed-off-by: Chris Packham <judge.packham@gmail.com>

Series-to: uboot
Series-to: net
Series-cc: hannah@marvell.com
Series-cc: jp.tosoni@acksys.fr
Series-cc: Angga <hermin.anggawijaya@alliedtelesis.co.nz>
Series-cc: reed@fb.com
Series-cc: luka.kovacica@gmail.com
Series-prefix: RFC
Series-version: 3

Cover-letter:
IPv6 support

This series adds basic IPv6 support to U-boot. It is a reboot of my
earlier work on this[1]. This latest version is itself a reboot of work
that was last posted to the list in late 2015. Right now I've just
rebased against master so functionally it is little different to what
was last posted.

Most of this is ported from Allied Telesis' additions to u-boot[2].
(Note that I am employed by Allied Telesis[3]). The hard work was done
some time ago by Angga, I've cleaned it up and made some improvements
with the hope of getting it accepted upstream.

A few open issues

1) rxhand_f currently takes an struct in_addr. TFTP doesn't use this
(I haven't looked at other users). To support V6 this may need to be a
new union, a void * with some kind of flag or nothing if no rxhandler
actually cares. It has been suggested that this parameter be removed and
any users that care can re-parse the packet.

2) Unit tests. This code needs them. The testing so-far has been ad-hoc
manual testing using qemu-x86.

3) Fancy v6 feature xyz. There are a lot of things that _could_ be
implemented on top of this (DHCPV6 and SLAAC are two obvious things that
stick out). For now I want to concentrate on getting the core code
stable and accepted, if anyone else wants to work on either of those
features that'd be great.

--
[1] - http://thread.gmane.org/gmane.comp.boot-loaders.u-boot/151390
[2] - http://www.alliedtelesis.co.nz/support/gpl/other.html
[3] - some of this has been done on work time, other parts have been
      done in my personal time. Since I'm subscribed to the list using
      my gmail account I've signed using that.

END